+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#1639** : Default index volume group property changes.
 
+* Issue **#1636** : Fix data retention deletion using wrong action for rules.
+
 * Issue **#1280** : Fix creation of default index volumes.
+
 
 ## [v7.0-beta.36] - 2020-06-02